src/test/java/com/example/Account/controller/AccountControllerTest.java
package com.example.Account.controller;

import com.example.Account.domain.Account;
import com.example.Account.type.AccountStatus;
import com.example.Account.service.AccountService;
import com.example.Account.service.RedisTestService;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.test.web.servlet.MockMvc;

import static org.mockito.ArgumentMatchers.anyLong;
import static org.mockito.BDDMockito.given;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;

@WebMvcTest(AccountController.class) // test하고자 하는 컨트롤러를 명시해줌.
class AccountControllerTest {
    // Mock이지만 Bean으로 자동 등록되어서 위의 컨트롤러에 자동으로 주입된다.
    // @MockBean
    private AccountService accountService;
    @MockBean
    private RedisTestService redisTestService;
    @Autowired
    private MockMvc mockMvc;
    @Test
    void successGetAccount() throws Exception {
        // given
        given(accountService.getAccount(anyLong()))
                .willReturn(Account.builder()
                        .accountNumber("3456")
                        .accountStatus(AccountStatus.IN_USE)
                        .build());
        // when
        // then
        mockMvc.perform(get("/account/876"))
                .andDo(print())
                .andExpect(jsonPath("$.accountNumber").value("3456"))
                .andExpect(jsonPath("$.accountStatus").value("IN_USE"))
                .andExpect(status().isOk());
    }
}